requirements.txt
torch
torchvision
pillow
matplotlib
pandas
scikit-learn
tqdm